Fix compile errors caused by refactoring
