+++ src/main/java/de/tum/www1/orion/vcs/OrionGitUtil.kt
@@ -20,7 +20,7 @@ import com.intellij.openapi.vcs.changes.VcsDirtyScopeManager
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
-import de.tum.www1.orion.bridge.ArtemisBridge
+import de.tum.www1.orion.messaging.OrionIntellijStateNotifier
 import de.tum.www1.orion.util.OrionFileUtils
 import de.tum.www1.orion.util.OrionSettingsProvider
 import de.tum.www1.orion.util.invokeOnEDTAndWait
@@ -66,7 +66,7 @@ object OrionGitUtil {
             private var parent: VirtualFile? = null
 
             override fun run(indicator: ProgressIndicator) {
-                ServiceManager.getService(project, ArtemisBridge::class.java).isCloning(true)
+                project.messageBus.syncPublisher(OrionIntellijStateNotifier.INTELLIJ_STATE_TOPIC).isCloning(true)
                 indicator.isIndeterminate = true
                 val lfs = LocalFileSystem.getInstance()
                 parent = lfs.findFileByIoFile(File(baseDir))
@@ -90,13 +90,13 @@ object OrionGitUtil {
                     directoryCheckedOut(File(baseDir, clonePath), GitVcs.getKey())
                     checkoutCompleted()
                 }
-                ServiceManager.getService(project, ArtemisBridge::class.java).isCloning(false)
+                project.messageBus.syncPublisher(OrionIntellijStateNotifier.INTELLIJ_STATE_TOPIC).isCloning(true)
                 andThen?.invoke()
             }
 
             override fun onError(error: Exception) {
                 super.onError(error)
-                ServiceManager.getService(project, ArtemisBridge::class.java).isCloning(false)
+                project.messageBus.syncPublisher(OrionIntellijStateNotifier.INTELLIJ_STATE_TOPIC).isCloning(false)
             }
         }.queue()
     }
